Add files via upload
Use this simple Python script to append anything you want to the beginning of a file name but also sequentially number the files as well.  Useful for labeling files prior to archiving to keep them organized for future reference.  Can work with any file type as long as the correct directory is used (enter between the single quotes after os.chdir) and a file extension type is supplied.